+++ statistics/statistics-api/src/main/java/com/duckduckgo/app/statistics/store/OfflinePixelCountDataStore.kt
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2022 DuckDuckGo
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.duckduckgo.app.statistics.store
+
+interface OfflinePixelCountDataStore {
+    var applicationCrashCount: Int
+    var webRendererGoneCrashCount: Int
+    var webRendererGoneKilledCount: Int
+    var cookieDatabaseNotFoundCount: Int
+    var cookieDatabaseOpenErrorCount: Int
+    var cookieDatabaseCorruptedCount: Int
+    var cookieDatabaseDeleteErrorCount: Int
+}
